// src/world/regions/generated/ocean.mask.ts
export const CELL_W = 8;
export const CELL_H = 11;
// ...
export const ANCHORS: Record<string, { x: number; y: number }> = {
  harbor: { x: 278, y: 241 },
  dock: { x: 292, y: 276 },
  spawn: { x: 348, y: 352 },
  school_sea_1: { x: 396, y: 352 },
  school_sea_2: { x: 468, y: 479 },
  school_sea_3: { x: 852, y: 662 },
  school_deep_1: { x: 667, y: 625 },
  school_deep_2: { x: 706, y: 659 },
  label_pacific: { x: 816, y: 338 },
  label_south: { x: 312, y: 662 },
};